gpio: add basic gpiod based gpio resource support
This adds initial support for gpio resources based on gpiod, instead of the
legacy sysfs interface.

This support is far form perfect of complete.
* ATM, '.get" is not suppported as using the command 'gpioget' would put
  the GPIO back in input mode.
* As stated in 'gpioset' help, the behavior of a gpio is undefined after
  'gpioset' exit. It works for an FTDI chip but it is should be considered
  fragile and unreliable.

Upgrading the gpiod resource driver to something better should be easy once
dbus support lands in libgpiod. This is still WIP ATM and may take a while
to land in major distros.

Until then, this is better than nothing.

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>