client: allow ssh commands to providers
Use the information provided by Remote Provider to create an ssh
connection.

This allow to use commands such as 'scp' or 'rsync' with the name of a
Provider. It allows the end user to place data on a provider serving a
'place' which worrying where that provider or how it is accessed. This is
helpful with workflows involving TFTP or NFS.

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>