exporter: allow proxy configuration
The exporter runs in isolated mode, the proxy default to the fqdn of the
host. That may work in most cases, but not always.

For example the hostname provided is like 'user@host.org', the information
about the user login gets lost and the isolated mode does not work.

With this change, the default proxy is still the host fqdn but it can
be changed with a command line argument

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>