provider: tftp: copy file to the target directory
The TFTP provider copies the file to the labgrid cache then adds
a symlinks to it in the TFTP exported directory.

That does not work if the deamon chroots to it own diretory
and the labgrid directory is not under it.

Override the staging function of the TFTP driver to actually
copy the files to TFTP directory

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>